uggly hack but recipe list seems to work
